MASSIVE UPDATE
Restructured the classes: all classes now have the base class MachineModule. Seperated the Grabber and Piston class: Piston is now the base class of Grabber.

the wait() funtion is now apart of the MachineModule class.

the only places the digitalWrite() functions occur are in the maintain() functions in evry class